+++ scripts/OIMS2fits.py
@@ -1,13 +1,7 @@
 #!/usr/bin/env python3
 
 from lsl.common.mcs import mjdmpm_to_datetime
-from lsl.sim.beam import beam_response
-from lsl.misc import parser as aph
-from scipy.interpolate import interp1d
 from astropy.io import fits as astrofits
-from astropy.coordinates import AltAz, EarthLocation, SkyCoord
-from astropy.wcs.utils import pixel_to_skycoord
-from astropy.time import Time
 from datetime import datetime, timedelta
 import os
 import sys
@@ -15,48 +9,11 @@ import numpy
 import argparse
 
 from lsl_toolkits.OrvilleImage import OrvilleImageDB
+
 from lsl_toolkits.OrvilleImage.wcs import WCS
 
+from lsl_toolkits.OrvilleImage.pbtools import pbcorroims
 
-def calcbeamprops(az,alt,header,freq):
-
-    # az and alt need to be the same shape as the image we will correct
-    
-    polarpatterns = []
-    polarpatterns.append(beam_response('empirical', 'XX', az, alt, frequency=freq))
-    polarpatterns.append(beam_response('empirical', 'YY', az, alt, frequency=freq))
-    
-    return polarpatterns[0], polarpatterns[1]
-
-def pbcorroims(header,imSize,chan,station):
-    pScale = header['pixel_size']
-    sRad   = 360.0/pScale/numpy.pi / 2
-    w = WCS.from_orville_header(header)
-    w = w.dropaxis(-1).dropaxis(-1)
-    x = numpy.arange(imSize) - 0.5
-    y = numpy.arange(imSize) - 0.5
-    x,y = numpy.meshgrid(x,y)
-    maskpix  = ((x-imSize/2.0)**2 + (y-imSize/2.0)**2) > ((0.98*sRad)**2)
-    x[maskpix] = imSize/2
-    y[maskpix] = imSize/2
-    sc = pixel_to_skycoord(x, y, wcs=w, mode='wcs')
-    # Need date and location for converting to altaz
-    if station == b'LWASV':
-        site = EarthLocation.from_geodetic(-106.885783, 34.348358, height=1477.8) 
-    elif station == b'LWANA':
-        site = EarthLocation.from_geodetic(-107.640, 34.247, height=2134)
-    time = Time(header['start_time'], header['int_len']/2, format='mjd', scale='utc')
-    aa = AltAz(location=site, obstime=time)
-    myaltaz = sc.transform_to(aa)
-    alt = myaltaz.alt.deg
-    az = myaltaz.az.deg
-    # Keep alt between 0 and 90, adjust az accordingly
-    negalt = alt < 0
-    alt[negalt] *= -1
-    az[negalt] += 180
-    freq = int(header['start_freq'])  + (chan*header['bandwidth'])
-    XX,YY = calcbeamprops(az,alt,header,freq)
-    return XX,YY
 
 def main(args):
     for filename in args.filename:
